Cart Screen Changes, Get Cart Details API Integrate, Add Cart in Product API Integrate

+++ food_delivery/lib/models/food_campaign_model/food_campaign_model.dart
@@ -252,7 +252,7 @@ class Restaurant {
   String tax;
   String zone;
   int numberOfReviews;
-  double rating;
+  int rating;
   List<String> campaignjoin;
 
   factory Restaurant.fromJson(Map<String, dynamic> json) => Restaurant(
@@ -285,7 +285,7 @@ class Restaurant {
     tax: json["Tax"] ?? "",
     zone: json["Zone"] ?? "",
     numberOfReviews: json["NumberOfReviews"] ?? 0,
-    rating: json["Rating"] ?? 0.0,
+    rating: json["Rating"] ?? 0,
     campaignjoin: List<String>.from(json["campaignjoin"].map((x) => x) ?? {}),
   );
